create,update,make offer,make order 27.05

+++ src/main/java/bg/softuni/lightthedeal/web/DTO/OfferMaterialLine.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 
 public record OfferMaterialLine(
         UUID materialId,
-        BigDecimal quantity
+        Double quantity
 ) {
 }
